CarbonFootPrintChatbot/requirements.txt
pandas
langchain
langchain_ibm
langchain_community
ibm_watson_machine_learning
ibm_watsonx_ai
faiss_cpu
sentence_transformers
docx2txt
chardet
python-dotenv
flask
